Fixed chunk position hashing so that all chunks have unique entries. Added x64 bit compile support

+++ MinecraftOpenGL/Source/ThreadPool.h
@@ -25,7 +25,7 @@ struct ChunkAction
 		type(type), chunk(chunk) {};
 
 	ActionType type = Rebuild;
-	unsigned int timestamp;
+	unsigned long long timestamp;
 
 	Chunk* chunk;
 	ChunkBuilder* chunkBuilder;
